requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# NLP and ML
spacy==3.7.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# File Processing
pdfplumber==0.10.3
python-docx==1.1.0
pypandoc==1.12

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
